Successful signup now redirects front end to "/" route in react router
